Feature routes refactoring cleaner (#425)
* break up routes into different files, add some middleware to cut out redundant code

* fix bug which causes 404 to fail

* ADD HELLA TESTS!!!
