+++ locales/es.json
@@ -376,5 +376,7 @@
 	"Planning Documents": "Planning Documents",
 	"Summary Documents": "Summary Documents",
 	"Reflection Reminders": "Reflection Reminders",
-	"CEU Documents": "CEU Documents"
+	"CEU Documents": "CEU Documents",
+	"New User Signups": "New User Signups",
+	"SOLEs Growth": "SOLEs Growth"
 }
