+++ server.js
@@ -50,12 +50,9 @@ app.use('/admin', require('./routes/router-admin.js')); //admin routes, auth-req
 app.use('/slackbot', require('./routes/router-slackbot.js')); //slackbot routes, unauth'ed
 app.use('/', require('./routes/router-ring-colombia.js')); //colombia ring routes
 app.use('/profile', require('./routes/router-profile.js'));
+app.use('/soles', require('./routes/router-soles.js'));
+app.use('/questions', require('./routes/router-questions.js'));
 app.use('/rings', require('./routes/router-rings.js'));
-
-// const ringsRouter = require('./routes/router-rings.js');
-// ringsRouter.use(middlewares.isAuth);
-// app.use('/rings', ringsRouter);
-
 
 // serve static content
 app.use(express.static(path.join(__dirname, 'public')));
