+++ tests/tests.js
@@ -4,64 +4,83 @@ const server        = require('../server.js'),
       should        = chai.should(),
       Parse         = require('parse/node'),
       soleConfig    = require('../sole-config.js');
-      sessionToken  = process.env.SESSION_TOKEN; //add sessionToken to environmental variables
+      sessionToken  = process.env.TEST_SESSION_TOKEN; //add sessionToken to environmental variables
 
 chai.use(chaiHttp);
 
-const routes = [
-  '/home',
-  '/history',
-  '/how',
-  '/terms-of-use',
-  '/privacy',
-  '/verify-email-success',
-  '/verify-email-failure',
-  '/resources',
-  '/profile',
-  '/profile/complete',
-  '/soles',
-  // '/soles/:id',
-  // '/soles/:id/download-plan',
-  // '/soles/:id/download-summary',
-  // '/soles/:id/copy',
-  // '/soles/:id/plan',
-  // '/soles/:id/delete',
-  // '/soles/:id/reflect',
-  // '/soles/reflect',
-  '/questions',
-  '/questions/mine',
-  '/questions/new',
-  // '/questions/:id',
-  // '/questions/:id/favorite',
-  // '/questions/:id/delete-tag/:rdn',
-  // '/questions/:id/approve',
-  // '/questions/:id/reject',
-  // '/soles/:id/approve',
-  // '/soles/:id/reject',
-  '/error'
-];
 
-describe('Test if all routes load', function() {
-  it('GET /. It should show login page', (done) => {
-    chai.request(server)
-      .get('/')
-      .end((err, res) => {
-        res.should.have.status(200);
-        done();
-      });
+const routes = {
+  prelogin: [
+    "/",
+    "/register",
+    "/logout",
+    // "/google-login-error",
+    "/login",
+    "/terms-of-use",
+    "/privacy",
+    "/verify-email-success",
+    "/verify-email-failure",
+    "/colombia",
+    "/colombia/register",
+    "/colombia/login"
+  ],
+  postlogin: [
+    "/home",
+    "/history",
+    "/how",
+    "/resources",
+    "/random-picture",
+    "/admin",
+    "/admin/pending-soles",
+    "/admin/browse-soles",
+    "/admin/browse-users",
+    "/admin/events",
+    "/admin/pa",
+    "/profile",
+    "/profile/subscriptions",
+    "/profile/about-me",
+    "/profile/complete",
+    "/soles",
+    "/soles/plan",
+    // "/soles/:id",
+    // "/soles/:id/download-plan",
+    // "/soles/:id/download-summary",
+    // "/soles/:id/copy",
+    // "/soles/:id/plan",
+    // "/soles/:id/delete",
+    // "/soles/:id/reflect",
+    "/questions",
+    // "/questions/mine",
+    "/questions/new",
+    // "/questions/:id",
+    // "/questions/:id/favorite",
+    // "/questions/:id/delete-tag/:rdn",
+    "/rings",
+    // "/rings/:id",
+    // "/rings/:id/educators-growth",
+    // "/rings/:id/educators-by-grade",
+    // "/rings/:id/soles-by-grade",
+    // "/rings/:id/soles-growth",
+    // "/rings/:id/educators-by-subject",
+    // "/rings/:id/soles-by-subject"
+  ]
+};
+// routes special cases: (must be post requests, maybe with param)
+// "/users-range",
+// "/users-today",
+// "/users-range-detail",
+
+describe('Load all pre-login routesload', function() {
+  routes.prelogin.forEach(route => {
+    it('GET' + route + '. Does page load?', (done) => {
+      chai.request(server)
+        .get(route)
+        .end((err, res) => {
+          res.should.have.status(200);
+          done();
+        });
+    }).timeout(10000);
   });
-
-  // routes.forEach(route => {
-  //   it('GET' + route + '. It should show page', (done) => {
-  //     chai.request(server)
-  //       .get(route)
-  //       .set('Cookie', 'sessionToken=' + sessionToken +' ;')
-  //       .end((err, res) => {
-  //         res.should.have.status(200);
-  //         done();
-  //       });
-  //   }).timeout(10000);
-  // });
 
   it('GET /lolwutno. It should show 404', (done) => {
     chai.request(server)
@@ -73,5 +92,20 @@ describe('Test if all routes load', function() {
   }).timeout(10000);
 
 });
+
+// describe('Test if all post-login routes load', function() {
+//   routes.postlogin.forEach(route => {
+//     it('GET' + route + '. Does page load?', (done) => {
+//       chai.request(server)
+//         .get(route)
+//         .set('Cookie', 'sessionToken=' + sessionToken +' ;')
+//         .end((err, res) => {
+//           res.should.have.status(200);
+//           done();
+//         });
+//     }).timeout(10000);
+//   });
+//
+// });
